separate glad and terrai into 2 routers
